utility/pyproject.toml
[tool.poetry]
name = "utility"
version = "0.1.0"
description = ""
authors = ["Javkhaa <javkhaa@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
pandas = "^2.2.0"
matplotlib = "^3.8.3"
opencv-python = "^4.9.0.80"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
